data: Language: chop withStrings
