+++ package.json
@@ -23,7 +23,6 @@
         "#composite/things/artwork": "./src/data/composite/things/artwork/index.js",
         "#composite/things/content": "./src/data/composite/things/content/index.js",
         "#composite/things/contribution": "./src/data/composite/things/contribution/index.js",
-        "#composite/things/language": "./src/data/composite/things/language/index.js",
         "#composite/things/track": "./src/data/composite/things/track/index.js",
         "#content-dependencies": "./src/content/dependencies/index.js",
         "#content-function": "./src/content-function.js",
