--- src/data/composite/things/language/index.js
@@ -1,1 +1,0 @@
-export {default as withStrings} from './withStrings.js';
